merger to main

+++ hobbyhobby-community/hobbyhobby-domain/src/main/java/org/v1/implementation/community/UserCommunityRemover.java
@@ -8,7 +8,7 @@ import org.v1.repository.UserCommunityRepository;
 @AllArgsConstructor
 public class UserCommunityRemover {
     private final UserCommunityRepository userCommunityRepository;
-    public void removeUserFromCommunity(Long userId, Integer communityId) {
+    public void removeUserFromCommunity(Long userId, Long communityId) {
         userCommunityRepository.removeUserFromCommunity(userId, communityId);
     }
 }
